# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the kidzzle project.",
        "title": "kidzzle API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/assessments/{assessmentType}/training": {
            "get": {
                "description": "Get training methods for a given assessment type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assessment"
                ],
                "summary": "Get assessment training methods",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Assessment Type ID",
                        "name": "assessmentType",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Training methods retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/domain.GetAssessmentTrainingMethodsResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/domain.ValidationError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponseExample401"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponseExample500"
                        }
                    }
                }
            }
        },
        "/assessments/{kidId}/{ageRageId}/{assessmentTypeId}/result": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get assessment result data for a given kid, age range, and assessment type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assessment"
                ],
                "summary": "Get assessment result data by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access Token",
                        "name": "accessToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Kid ID",
                        "name": "kidId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Age Range ID",
                        "name": "ageRageId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Assessment Type ID",
                        "name": "assessmentTypeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Assessment result data retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/domain.GetAssessmentResultDataByIdResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/domain.ValidationError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponseExample401"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponseExample500"
                        }
                    }
                }
            }
        },
...
        "/kids/{kidId}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the data of a kid by their ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kid"
                ],
                "summary": "Update kid data by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access Token",
                        "name": "accessToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Kid ID",
                        "name": "kidId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Kid Data Update Request",
                        "name": "kidData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateKidDataByIdRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kid data updated successfully",
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateKidDataByIdResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/domain.ValidationError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponseExample401"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponseExample500"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete kid data by given kid ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kid"
                ],
                "summary": "Delete kid data by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access Token",
                        "name": "accessToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Kid ID",
                        "name": "kidId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kid data deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/domain.DeleteKidDataByIdResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/domain.ValidationError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponseExample401"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponseExample500"
                        }
                    }
                }
            }
        },
...
        "/promotions/body-part": {
            "get": {
                "description": "Retrieve the promotion body part details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Promotion"
                ],
                "summary": "Get promotion body part",
                "responses": {
                    "200": {
                        "description": "Promotion body part retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/domain.GetPromotionBodyPartResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/domain.ValidationError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponseExample500"
                        }
                    }
                }
            }
        },
        "/promotions/color": {
            "get": {
                "description": "Retrieve the promotion color",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Promotion"
                ],
                "summary": "Get promotion color",
                "responses": {
                    "200": {
                        "description": "Promotion color retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/domain.GetPromotionColorResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/domain.ValidationError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponseExample500"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.DeleteKidDataByIdResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "domain.ErrorResponseExample401": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 401
                },
                "message": {
                    "type": "string",
                    "example": "Unauthorized"
                }
            }
        },
        "domain.ErrorResponseExample500": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "message": {
                    "type": "string",
                    "example": "Internal Server Error"
                }
            }
        },
...
        "domain.GetAssessmentResultDataByIdResponse": {
            "type": "object",
            "properties": {
                "age_range_id": {
                    "type": "string",
                    "example": "AGER_1"
                },
                "assessment_question_id": {
                    "type": "string",
                    "example": "3afe2697-2b02-46d4-a9c3-efd0bdd9e014"
                },
                "assessment_result_id": {
                    "type": "string",
                    "example": "7d7bc87f-0faf-43f9-9af9-1a09166db993"
                },
                "assessment_type_id": {
                    "type": "string",
                    "example": "ASSMTT_1"
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-03-09T00:00:00Z"
                },
                "is_passed": {
                    "type": "boolean",
                    "example": true
                },
                "kid_id": {
                    "type": "string",
                    "example": "8457529e-01fb-41d0-82b1-9618be667862"
                }
            }
        },
        "domain.GetAssessmentTrainingMethodsResponse": {
            "type": "object",
            "properties": {
                "assessment_no": {
                    "type": "string",
                    "example": "1"
                },
                "assessment_question_id": {
                    "type": "string",
                    "example": "088e7d6b-faa8-11ef-870e-0242ac120002"
                },
                "assessment_type_id": {
                    "type": "string",
                    "example": "ASSMTT_1"
                },
                "training_methods_id": {
                    "type": "string",
                    "example": "088e7d6b-faa8-11ef-870e-0242ac120002"
                },
                "training_required_tools": {
                    "type": "string",
                    "example": "Training Required Tools"
                },
                "training_text": {
                    "type": "string",
                    "example": "Training Text"
                }
            }
        },
...
        "domain.GetPromotionBodyPartResponse": {
            "type": "object",
            "properties": {
                "body_id": {
                    "type": "string",
                    "example": "bd5f5f6f-fd05-11ef-870e-0242ac120002"
                },
                "body_image_url": {
                    "type": "string",
                    "example": "url_link"
                },
                "body_name": {
                    "type": "string",
                    "example": "สีเขียว (Green)"
                },
                "body_sound_file_url": {
                    "type": "string",
                    "example": "url_link"
                }
            }
        },
        "domain.GetPromotionColorResponse": {
            "type": "object",
            "properties": {
                "color_id": {
                    "type": "string",
                    "example": "bd5f5f6f-fd05-11ef-870e-0242ac120002"
                },
                "color_image_url": {
                    "type": "string",
                    "example": "url_link"
                },
                "color_name": {
                    "type": "string",
                    "example": "สีน้ำตาล (Brown)"
                },
                "color_shape": {
                    "type": "string",
                    "example": "สี่เหลี่ยมจัตุรัส (Square)"
                },
                "color_sound_file_url": {
                    "type": "string",
                    "example": "url_link"
                }
            }
        },
        "domain.UpdateKidDataByIdRequest": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "kidId": {
                    "type": "string"
                },
                "kid_birth_weight": {
                    "type": "string"
                },
                "kid_birthday": {
                    "type": "string"
                },
                "kid_blood_type": {
                    "type": "string"
                },
                "kid_body_length": {
                    "type": "string"
                },
                "kid_congenital_disease": {
                    "type": "string"
                },
                "kid_gender": {
                    "type": "string"
                },
                "kid_gestational_age": {
                    "type": "string"
                },
                "kid_name": {
                    "type": "string"
                },
                "kid_oxygen": {
                    "type": "string"
                },
                "pregnant_id": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "domain.UpdateKidDataByIdResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "domain.ValidationError": {
            "type": "object",
            "properties": {
                "errors": {
                    "$ref": "#/definitions/domain.ValidationErrorResponse"
                }
            }
        },
        "domain.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "auth_error": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "incorrect": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "user_error": {
                    "type": "string"
                }
            }
        }
    }
}